Create agnostic module for do account operation
