+++ lib/elixir_pay/accounts/withdraw.ex
@@ -1,46 +1,11 @@
 defmodule ElixirPay.Accounts.Withdraw do
-  alias Ecto.Multi
+  alias ElixirPay.Accounts.Operation
+  alias ElixirPay.Repo
 
-  alias ElixirPay.{Account, Repo}
-
-  def call(%{"id" => id, "value" => value}) do
-    Multi.new()
-    |> Multi.run(:account, fn repo, _changes -> get_account(repo, id) end)
-    |> Multi.run(:update_balance, fn repo, %{account: account} ->
-      update_balance(repo, account, value)
-    end)
+  def call(params) do
+    params
+    |> Operation.call(:withdraw)
     |> run_transaction()
-  end
-
-  defp get_account(repo, id) do
-    case repo.get(Account, id) do
-      nil -> {:error, "Account, not found!"}
-      account -> {:ok, account}
-    end
-  end
-
-  defp update_balance(repo, account, value) do
-    account
-    |> sum_values(value)
-    |> update_account(repo, account)
-  end
-
-  defp sum_values(%Account{balance: balance}, value) do
-    value
-    |> Decimal.cast()
-    |> handle_cast(balance)
-  end
-
-  defp handle_cast({:ok, value}, balance), do: Decimal.sub(balance, value)
-  defp handle_cast(:error, _balance), do: {:error, "Invalid deposit value!"}
-
-  defp update_account({:error, _reason} = error, _repo, _account), do: error
-  defp update_account(value, repo, account) do
-    params = %{balance: value}
-
-    account
-    |> Account.changeset(params)
-    |> repo.update()
   end
 
   defp run_transaction(multi) do
